# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/": {
            "get": {
                "produces": [
                    "text/html"
                ],
                "responses": {
                    "200": {
                        "description": "success request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/contracts.ErrorModel"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/contracts.ErrorModel"
                        }
                    }
                }
            }
        },
        "/update": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "metric",
                        "name": "metric",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contracts.Metric"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/contracts.ErrorModel"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/contracts.ErrorModel"
                        }
                    }
                }
            }
        },
        "/update/{type}/{name}/{value}": {
            "post": {
                "produces": [
                    "plain/text",
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Metric type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Metric name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Metric value",
                        "name": "value",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/contracts.ErrorModel"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/contracts.ErrorModel"
                        }
                    }
                }
            }
        },
        "/updates": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "metric array",
                        "name": "metrics",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/contracts.Metric"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/contracts.ErrorModel"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/contracts.ErrorModel"
                        }
                    }
                }
            }
        },
        "/value": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "metric",
                        "name": "metric",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contracts.Metric"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success request",
                        "schema": {
                            "$ref": "#/definitions/contracts.Metric"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/contracts.ErrorModel"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/contracts.ErrorModel"
                        }
                    }
                }
            }
        },
        "/value/{type}/{name}": {
            "get": {
                "produces": [
                    "plain/text",
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Metric type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Metric name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "contracts.ErrorModel": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "contracts.Metric": {
            "type": "object",
            "properties": {
                "delta": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        }
    }
}